Refactor: Move provider-related files to new `provider` package.
This commit reorganizes the project structure by moving all files related to currency exchange providers into a new dedicated package `com.currency.rateman.provider`.

This includes:
- Data models (`Provider.kt`)
- DAO (`ProviderDao.kt`)
- Repository interfaces and implementations (`ProviderRepository.kt`, `ProviderRepositoryImpl.kt`)
- ViewModels (`ProviderViewModel.kt`)
- UI components and screens (`ProviderItem.kt`, `ProvidersList.kt`, `ProviderDetailsScreen.kt`)
- Utility files (`ProviderConstants.kt`)

Imports and references have been updated accordingly across the codebase.
