+++ app/src/main/java/com/currency/rateman/api/kurzyCz/ProviderConverter.kt
@@ -5,7 +5,7 @@ import com.currency.rateman.api.kurzyCz.ProviderTypeClassifier
 import com.currency.rateman.data.db.entity.CurrencyRateEntity
 import com.currency.rateman.data.db.entity.ProviderEntity
 import com.currency.rateman.data.model.CurrencyRate
-import com.currency.rateman.data.model.Provider
+import com.currency.rateman.provider.data.model.Provider
 import com.currency.rateman.data.model.enums.CurrencyCode
 import com.currency.rateman.data.model.enums.ProviderType
 import java.time.LocalDate
