+++ app/src/main/java/com/currency/rateman/api/kurzyCz/ProviderTypeClassifier.kt
@@ -1,7 +1,7 @@
 package com.currency.rateman.api.kurzyCz
 
 import com.currency.rateman.data.model.enums.ProviderType
-import com.currency.rateman.utils.ProviderConstants
+import com.currency.rateman.provider.utils.ProviderConstants
 
 object ProviderTypeClassifier {
     fun determineProviderType(bankName: String): ProviderType {
