+++ app/src/main/java/com/currency/rateman/provider/domain/repository/ProviderRepository.kt
@@ -1,7 +1,7 @@
-package com.currency.rateman.data.repository
+package com.currency.rateman.provider.domain.repository
 
 import com.currency.rateman.api.kurzyCz.ProviderAPI
-import com.currency.rateman.data.model.Provider
+import com.currency.rateman.provider.data.model.Provider
 import kotlinx.coroutines.flow.Flow
 
 interface ProviderRepository {
